Update contest status pojo according to api endpoint
